chore(rsc): Refactor: Split rscWorker into worker and renderer
